Rename the existing test because it was misleading. Create a unit test to verify the use of ExitableDirectoryReader.
